Move modal helper to its own file
