+++ app/helpers/application_helper.rb
@@ -9,19 +9,4 @@ module ApplicationHelper
       render partial: "shared/alert", locals: { message: message }
     end
   end
-
-  def modal_link_to(text, path, **options)
-    id = options[:id] || path
-    turbo_frame_tag id do
-      link_to text, path, **options
-    end
-  end
-
-  def modal_content_for(id, parent, &block)
-    render partial: 'shared/modal', locals: {
-      id: id,
-      parent: parent,
-      block: block
-    }
-  end
 end
